LayoutTests/fast/forms/datalist/datalist-show-hide-expected.txt

CLICK
PASS showingList is "true"
BLUR
PASS showingList is "false"
PASS successfullyParsed is true

TEST COMPLETE
